Derive Macro redefined_derive::Redefined

#[derive(Redefined)]
{
    // Attributes available to this derive:
    #[redefined]
    #[redefined_attr]
}

DERIVE MACRO

Container Source Attribute

WITH: #[redefined(<TYPE IDENTIFIER>)] - Only used to specify the source struct (i.e. the struct converting from)

Example:

    #[derive(Debug, Clone, PartialEq, Default, Redefined)]
    #[redefined(BasicStruct)]
    pub struct BasicStructA {
        pub val1: u64,
        pub val2: f64,
        pub val3: String,
    }

WITHOUT: #[redefined(<TYPE IDENTIFIER>)] - Omitted when the current type is the source type to a new type - The new type will be created with the same fields and attributes (except redefined attributes) - The new type’s name will be the same as this type’s name with a concatinated Redefined - The new type implements from/into this type via transmute - TODO: other methods besides transmute

Example:

    #[derive(Debug, Clone, PartialEq, Redefined)]
    pub struct GenericConstantStructA<const XVAL: usize> {
        pub p: u64,
        pub d: [i128; XVAL],
    }

    // GenericConstantStructARedefined is created

Container Attributes

#[redefined_attr(...)]

to_source = "..."

specifies a function to call when converting this type back into it’s source type

(i.e. for when fields are private)

from_source = "..."

specifies a function to call when converting the source type into this type

NOTE WHEN USING THIS: the source type is a variable called src (see first example)

(i.e. for when fields are private)

transmute

converts between the types using an unsafe transmute

conflicts with 1 and/or 2

derive(...)

specifies #[derive(...)] values for the new type

Examples:

to_source = '..' + from_source = '..'

    #[derive(Debug, Clone, PartialEq, Default, Redefined)]
    #[redefined(PrivateFieldStruct)]
    #[redefined_attr(
        to_source = "PrivateFieldStruct::new(self.p, self.d,self.vals)",
        from_source = "ToFromSourceFieldStructB::new(src)"
    )]
    pub struct ToFromSourceFieldStructB {
        pub p:    u64,
        pub d:    u64,
        pub vals: Vec<String>,
    }

transmute

    #[derive(Debug, Clone, PartialEq, Default, Redefined)]
    #[redefined(BasicStruct)]
    #[redefined_attr(transmute)]
    pub struct BasicStructA {
        pub val1: u64,
        pub val2: f64,
        pub val3: String,
    }

derive

    #[derive(Debug, Clone, PartialEq, Default, Redefined)]
    #[redefined_attr(derive(Debug, Clone, PartialEq, Default))]
    pub struct BasicStructA {
        pub val1: u64,
        pub val2: f64,
        pub val3: String,
    }

Field Attributes

#[redefined(...)]

These attributes go on over fields of the type

func = ".."

specifies a function to call when getting a field from the source struct

NOTE WHEN USING THIS: the source type is a variable called src (same as in from_source = '..' container attribute above)

field(...)

specifies a type to use when converting between target and source types

syntax: comma seperated list of (<SOURCE TYPE>, <TARGET TYPE>)

NOTE: only used when the container attribute redefined(...) is omitted

Sub-attributes:

field((<SOURCE TYPE>, same))

same can be used in place of the target type, when the target type’s identifier is the same as the source types

field((<SOURCE TYPE>, <TARGET TYPE>))

used when you want to convert the <SOURCE TYPE> to a custom <TARGET TYPE>

same_fields

all non-primitive types are converted to the predefined types by default, so passing this in will use the same value in the field for the redefined type

akin to using field((<SOURCE TYPE 1>, same), (<SOURCE TYPE 2>, same), ..)

if a type (owned by you) needs to implement RedefinedConvert, call self_convert_redefined(<TYPE>) to implement the trait for itself

#Examples:

func = '..'

    #[derive(Debug, Clone, PartialEq, Default, Redefined)]
    #[redefined(PrivateFieldStruct)]
    #[redefined_attr(to_source = "PrivateFieldStruct::new(self.p, self.d, self.vals)")]     
    pub struct NonPubFieldStructB {
        #[redefined(func = "src.get_p()")]
        pub p:    u64,
        pub d:    u64,
        pub vals: Vec<String>,
    }

field((.., ..)) + field((.., same))

    #[derive(Debug, Clone, PartialEq, Default, Redefined)]
    #[redefined_attr(derive(Debug, Clone, PartialEq))]
    pub struct ComplexStructAA<'a, 'b> {
        pub n:       i128,
        #[redefined(field((GenericLifetimeStructA, same), (BasicStructA, BasicStructARedefined)))]
        pub inner_a: Vec<(GenericLifetimeStructA<'a, 'b>, BasicStructA)>,
    }
    // the new type will have field:
    pub inner_a: Vec<(GenericLifetimeStructA<'a, 'b>, BasicStructARedefined)>,

/// same_fields

    #[derive(Debug, Clone, PartialEq, Default, Redefined)]
    #[redefined_attr(derive(Debug, Clone, PartialEq))]
    pub struct ComplexStructAA<'a, 'b> {
        pub n:       i128,
        #[redefined(same_fields)]
        pub inner_a: Vec<(GenericLifetimeStructA<'a, 'b>, BasicStructA)>,
    }
    // the new type will have field:
    pub inner_a: Vec<(GenericLifetimeStructA<'a, 'b>, BasicStructA)>,